合并代码

+++ ehealth-recipe-core/src/main/java/recipe/serviceprovider/drug/service/RemoteDrugService.java
@@ -57,7 +57,7 @@ public class RemoteDrugService extends BaseService<DrugListBean> implements IDru
             num = drugListDAO.getSpecifyNum(organId, drugType);
         }
 
-        return num;
+        return num.longValue();
     }
 
     @RpcService
